refactor(frontend): use dtos index import
