+++ frontend/__tests__/.server/domain/mappers/client-friendly-status.dto.mapper.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import type { ClientFriendlyStatusDto } from '~/.server/domain/dtos/client-friendly-status.dto';
+import type { ClientFriendlyStatusDto } from '~/.server/domain/dtos';
 import type { ClientFriendlyStatusEntity } from '~/.server/domain/entities/client-friendly-status.entity';
 import { ClientFriendlyStatusDtoMapperImpl } from '~/.server/domain/mappers';
