+++ frontend/__tests__/.server/domain/mappers/country.dto.mapper.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import type { CountryDto } from '~/.server/domain/dtos/country.dto';
+import type { CountryDto } from '~/.server/domain/dtos';
 import type { CountryEntity } from '~/.server/domain/entities/country.entity';
 import { CountryDtoMapperImpl } from '~/.server/domain/mappers';
