+++ frontend/__tests__/.server/domain/mappers/marital-status.dto.mapper.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
-import type { MaritalStatusDto } from '~/.server/domain/dtos/marital-status.dto';
+import type { MaritalStatusDto } from '~/.server/domain/dtos';
 import type { MaritalStatusEntity } from '~/.server/domain/entities/marital-status.entity';
 import { MaritalStatusDtoMapperImpl } from '~/.server/domain/mappers';
