+++ frontend/__tests__/.server/domain/services/marital-status.service.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
-import type { MaritalStatusDto } from '~/.server/domain/dtos/marital-status.dto';
+import type { MaritalStatusDto } from '~/.server/domain/dtos';
 import type { MaritalStatusDtoMapper } from '~/.server/domain/mappers';
 import type { MaritalStatusRepository } from '~/.server/domain/repositories/marital-status.repository';
 import { MaritalStatusServiceImpl } from '~/.server/domain/services/marital-status.service';
