+++ frontend/__tests__/.server/domain/services/province-territory-state.service.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
-import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos/province-territory-state.dto';
+import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos';
 import type { ProvinceTerritoryStateDtoMapper } from '~/.server/domain/mappers';
 import type { ProvinceTerritoryStateRepository } from '~/.server/domain/repositories/province-territory-state.repository';
 import { ProvinceTerritoryStateServiceImpl } from '~/.server/domain/services/province-territory-state.service';
