+++ frontend/app/.server/domain/mappers/preferred-language.dto.mapper.ts
@@ -2,7 +2,7 @@ import { inject, injectable } from 'inversify';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
-import type { PreferredLanguageDto } from '~/.server/domain/dtos/preferred-language.dto';
+import type { PreferredLanguageDto } from '~/.server/domain/dtos';
 import type { PreferredLanguageEntity } from '~/.server/domain/entities/preferred-language.entity';
 
 export interface PreferredLanguageDtoMapper {
