+++ frontend/app/.server/domain/mappers/province-territory-state.dto.mapper.ts
@@ -1,6 +1,6 @@
 import { injectable } from 'inversify';
 
-import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos/province-territory-state.dto';
+import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos';
 import type { ProvinceTerritoryStateEntity } from '~/.server/domain/entities/province-territory-state.entity';
 
 export interface ProvinceTerritoryStateDtoMapper {
