+++ frontend/app/.server/domain/services/country.service.ts
@@ -3,7 +3,7 @@ import moize from 'moize';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
-import type { CountryDto } from '~/.server/domain/dtos/country.dto';
+import type { CountryDto } from '~/.server/domain/dtos';
 import type { CountryDtoMapper } from '~/.server/domain/mappers';
 import type { CountryRepository } from '~/.server/domain/repositories/country.repository';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
