+++ frontend/app/.server/domain/services/province-territory-state.service.ts
@@ -3,7 +3,7 @@ import moize from 'moize';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
-import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos/province-territory-state.dto';
+import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos';
 import type { ProvinceTerritoryStateDtoMapper } from '~/.server/domain/mappers';
 import type { ProvinceTerritoryStateRepository } from '~/.server/domain/repositories/province-territory-state.repository';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
